+++ HISTORY.md
@@ -52,7 +52,8 @@
 ### 2021-11-08
 * objects: renamed some internal color methods
 * objects, store: renamed internal pen color channel cache
-* objects, blocks, threads: renamed internal pen accessor methods 
+* objects, blocks, threads: renamed internal pen accessor methods
+* objects, threads, store, extensions: switched pen color dimensions from HSV to HSL
 
 ### 2021-11-07
 * widgets, blocks, byob: allow block-instances to be dragged off from templates in the "hide blocks" dialog